ChangeElementTypeCommand extracted in the dedicated package
